Renaming z3_scheduler_new to z3_scheduler.
